Allow IGDB remasters and ports in game lookups

+++ backend/__tests__/setup.js
@@ -1,20 +1,30 @@
 const mongoose = require('mongoose')
 const { MongoMemoryServer } = require('mongodb-memory-server')
 
-let mongod
+if (process.env.SKIP_MONGO_MEMORY === 'true') {
+  beforeAll(async () => {})
+  afterEach(async () => {})
+  afterAll(async () => {
+    if (mongoose.connection.readyState !== 0) {
+      await mongoose.disconnect()
+    }
+  })
+} else {
+  let mongod
 
-beforeAll(async () => {
-  mongod = await MongoMemoryServer.create()
-  const uri = mongod.getUri()
-  await mongoose.connect(uri)
-})
+  beforeAll(async () => {
+    mongod = await MongoMemoryServer.create()
+    const uri = mongod.getUri()
+    await mongoose.connect(uri)
+  })
 
-afterEach(async () => {
-  const collections = mongoose.connection.collections
-  await Promise.all(Object.values(collections).map((col) => col.deleteMany({})))
-})
+  afterEach(async () => {
+    const collections = mongoose.connection.collections
+    await Promise.all(Object.values(collections).map((col) => col.deleteMany({})))
+  })
 
-afterAll(async () => {
-  await mongoose.disconnect()
-  await mongod.stop()
-})
+  afterAll(async () => {
+    await mongoose.disconnect()
+    await mongod.stop()
+  })
+}
